Fixed the CI test failure.
Root cause: the CI excerpt showed only MSBuild’s wrapper error, but the real xUnit failure was the gateway structure budget test: GatewayStructureTests.cs reported RuntimeInitializationExtensions.cs had grown to 299 lines, over the 252-line budget. The recent /loop gateway callback block caused the overflow.
