+++ src/OpenClaw.Gateway/Composition/RuntimeInitializationExtensions.cs
@@ -28,7 +28,6 @@ using OpenClaw.Gateway.Profiles;
 using OpenClaw.Gateway.Tools;
 using OpenClaw.Gateway.Pipeline;
 using OpenClaw.Plugins.TokenJuice;
-using OpenClaw.Core.Loops;
 
 namespace OpenClaw.Gateway.Composition;
 
@@ -64,7 +63,6 @@ internal static partial class RuntimeInitializationExtensions
         var services = ResolveRuntimeServices(app);
         RecordLegacyMafConfigNotice(app, services, startupLogger, startupNoticeSink);
         var providerSmokeRegistry = app.Services.GetRequiredService<ProviderSmokeRegistry>();
-
         var approvalService = app.Services.GetRequiredService<ToolApprovalService>();
         Telemetry.RegisterApprovalQueueGauge(() => approvalService.PendingCount);
         var blockedPluginIds = services.PluginHealth.GetBlockedPluginIds();
@@ -179,7 +177,6 @@ internal static partial class RuntimeInitializationExtensions
             services.ToolSandbox,
             interceptors);
         runtimeForLoadSkill = agentRuntime;
-
         if (agentRuntime is AgentRuntime concreteRuntime)
         {
             services.CommandProcessor.SetCompactCallback(async (session, ct) =>
@@ -189,51 +186,7 @@ internal static partial class RuntimeInitializationExtensions
             });
         }
 
-        // Wire /loop callback
-        services.CommandProcessor.SetLoopCallback(async (session, text, ct) =>
-        {
-            var scheduler = app.Services.GetRequiredService<ClawLoopScheduler>();
-            var cmd = LoopCommandParser.TryParse(text);
-            if (cmd is null || cmd.Action == LoopAction.Invalid)
-                return "Usage: /loop <interval> <prompt>  — e.g. /loop 5m check build status\n       /loop cancel  — cancel active loop\n       /loop status  — show loop status";
-
-            return cmd.Action switch
-            {
-                LoopAction.Cancel =>
-                    await CancelLoopAsync(scheduler, session.Id, ct),
-                LoopAction.Status =>
-                    await GetLoopStatusAsync(scheduler, session.Id, ct),
-                LoopAction.Schedule when cmd.Interval is not null && cmd.Prompt is not null =>
-                    await ScheduleLoopAsync(scheduler, session.Id, cmd.Interval, cmd.Prompt, ct),
-                _ => "Invalid /loop syntax."
-            };
-
-            static async Task<string> ScheduleLoopAsync(ClawLoopScheduler scheduler, string sessionId, string interval, string prompt, CancellationToken ct)
-            {
-                try
-                {
-                    var cron = ClawLoopScheduler.IntervalToCron(interval);
-                    await scheduler.ScheduleLoopAsync(sessionId, cron, prompt, ct);
-                    return $"Loop started — interval: {interval}, prompt: \"{prompt}\"";
-                }
-                catch (ArgumentException ex)
-                {
-                    return $"Error: {ex.Message}";
-                }
-            }
-
-            static async Task<string> CancelLoopAsync(ClawLoopScheduler scheduler, string sessionId, CancellationToken ct)
-            {
-                await scheduler.CancelLoopAsync(sessionId, ct);
-                return "Loop canceled.";
-            }
-
-            static async Task<string> GetLoopStatusAsync(ClawLoopScheduler scheduler, string sessionId, CancellationToken ct)
-            {
-                var status = await scheduler.GetLoopStatusAsync(sessionId, ct);
-                return status ?? "No active loop for this session.";
-            }
-        });
+        WireLoopCommandCallback(app, services);
 
         var middlewarePipeline = CreateMiddlewarePipeline(config, loggerFactory, services.ContractGovernance, services.SessionManager);
         var skillWatcher = new SkillWatcherService(
